Add restartable signing option to t_cose
This commit adds the option to execute the signing of a COSE object in
a restartable manner. This means that the function call executing the
signing may return before the signing of the COSE object is complete
with T_COSE_ERR_SIG_IN_PROGRESS. In this case the function needs to be
called again with the same parameters, until T_COSE_SUCCESS or an error
is returned. Restartable option is enabled to ECDSA, but not EdDSA.

This commit also updates the PSA, OpenSSL and test crypto adapters so
that they return T_COSE_ERR_UNSUPPORTED_RESTARTABLE_MODE.

Change-Id: Ifdb9145d6c538ffe1dd533d626e13e11c37cd64a